backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
opencv-python==4.8.1.78
mediapipe==0.10.21
numpy==1.26.4
pillow==10.2.0
protobuf>=4.25.3
